add warning switches
I am pleasantly surprised to see that I have basically zero warnings in
my code, despite never having used the compiler to check for them lol.

+++ external/stb_sprintf.h
@@ -591,7 +591,8 @@ STBSP__PUBLICDEF int STB_SPRINTF_DECORATE(vsprintfcb)(STBSP_SPRINTFCB *callback,
             s = (char *)"null";
          // get the length, limited to desired precision
          // always limit to ~0u chars since our counts are 32b
-         l = stbsp__strlen_limited(s, (pr >= 0) ? pr : ~0u);
+         // @Sol Cast pr to unsigned to silence compiler.
+         l = stbsp__strlen_limited(s, (pr >= 0) ? (uint)pr : ~0u);
          lead[0] = 0;
          tail[0] = 0;
          pr = 0;
